chats

+++ database/migrations/2024_08_09_225305_create_messages_table.php
@@ -10,20 +10,22 @@ return new class extends Migration
      * Run the migrations.
      */
 
-        public function up(): void
-        {
-            Schema::create('messages', function (Blueprint $table) {
-                $table->id();
-                $table->unsignedBigInteger('user_id');
-                $table->unsignedBigInteger('chat_id');
-   
-                $table->text('message');
-                $table->timestamps();
-
-                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-                $table->foreign('chat_id')->references('id')->on('chats')->onDelete('cascade');
-            });
-        }
+     public function up(): void
+     {
+         Schema::create('messages', function (Blueprint $table) {
+             $table->id();
+             $table->unsignedBigInteger('user_id');
+             $table->unsignedBigInteger('chat_id');
+             $table->text('message');
+             $table->timestamps();
+     
+             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+     
+             if (Schema::hasTable('chats')) {
+                 $table->foreign('chat_id')->references('id')->on('chats')->onDelete('cascade');
+             }
+         });
+     }
 
 
     /**
